+++ Constants.py
@@ -1,0 +1,37 @@
+class GNN_Methods:
+    Graph_SAINT = "G_SAINT"
+    RGCN = "RGCN"
+    ShaDowGNN = "ShaDowSAINT"
+    SeHGGN = "SeHGNN"
+    IBS = "IBS"
+    MorsE = "MorsE"
+    NodePiece = "NodePiece"
+    LHGNN = "LHGNN"
+    def __init__(self):
+        ""
+class KGE_Methods:
+    TransE = "TransE"
+    CompEx = "CompEx"
+    RotatE = "RotatE"
+    DistMult = "DistMult"
+    def __init__(self):
+        ""
+class GML_Operator_Types:
+    NodeClassification = "NC"
+    LinkPrediction = "LP"
+    GraphClassification = "GC"
+    def __init__(self):
+        ""
+class GML_Query_Types:
+    Inference = "Select"
+    Insert = "Insert"
+    Delete = "Delete"
+    def __init__(self):
+       ""
+class TOSG_Patterns:
+    d1h1 = "d1h1"
+    d1h2 = "d1h2"
+    d2h1 = "d2h1"
+    d2h2 = "d2h2"
+    def __init__(self):
+       ""
